# core-service/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.AchievementProgress:
    properties:
      current:
        type: integer
      granted_at:
        type: string
      key:
        type: string
      target:
        type: integer
    type: object
...
  domain.Freeze:
    properties:
      canceled_at:
        type: string
      created_at:
        type: string
      ends_at:
        type: string
      id:
        type: integer
      room_id:
        type: integer
      starts_at:
        type: string
      user_id:
        type: integer
    type: object
...
  domain.Stats:
    properties:
      best_streak:
        type: integer
      buddies:
        type: integer
      comments:
        type: integer
      early_workouts:
        type: integer
      late_workouts:
        type: integer
      rooms:
        type: integer
      total_workouts:
        type: integer
    type: object
  domain.User:
    properties:
      created_at:
        type: string
      first_name:
        type: string
      has_avatar:
        type: boolean
      id:
        type: integer
      photo_url:
        type: string
      rank:
        type: string
      status_emoji:
        type: string
      status_text:
        type: string
      theme:
        type: string
      username:
        type: string
    type: object
  httpapi.AuthTelegramRequest:
    properties:
      init_data:
        example: query_id=...&user=...&auth_date=...&hash=...
        type: string
    type: object
  httpapi.AuthTelegramResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/domain.User'
    type: object
  httpapi.BuddiesRequest:
    properties:
      user_ids:
        items:
          type: integer
        type: array
    type: object
...
  httpapi.ErrorResponse:
    properties:
      error:
        example: room not found
        type: string
    type: object
  httpapi.FreezeRequest:
    properties:
      ends_at:
        example: "2026-08-05"
        type: string
      starts_at:
        example: "2026-07-25"
        type: string
    type: object
...
  httpapi.MeResponse:
    properties:
      achievements:
        items:
          $ref: '#/definitions/domain.AchievementProgress'
        type: array
      best_streak:
        type: integer
      stats:
        $ref: '#/definitions/domain.Stats'
      user:
        $ref: '#/definitions/domain.User'
    type: object
...
  httpapi.UpdateMeRequest:
    properties:
      status_emoji:
        type: string
      status_text:
        example: На массе
        type: string
      theme:
        enum:
        - default
        - dark
        - neon
        example: dark
        type: string
    type: object
...
info:
  contact: {}
  description: 'Core service of BuddyGym: auth, users, rooms, rewards, checkin proxy.'
  title: BuddyGym Core API
  version: "1.0"
paths:
  /auth/telegram:
    post:
      consumes:
      - application/json
      description: Validates Mini App initData signature, upserts the user and issues
        a Bearer token for all other endpoints.
      parameters:
      - description: raw initData string
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.AuthTelegramRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.AuthTelegramResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Exchange Telegram initData for a JWT
      tags:
      - auth
...
  /checkins/{id}/buddies:
    post:
      consumes:
      - application/json
      description: Adds buddies to a checkin that is still being voted on. Only the
        author can tag, and buddies who are not members of the checkin room are ignored.
        The workout is credited to them once the room approves the checkin, so a rejected
        photo hands out nothing.
      parameters:
      - description: checkin id
        in: path
        name: id
        required: true
        type: string
      - description: tagged user ids
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.BuddiesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tag people who trained with you
      tags:
      - checkins
  /checkins/{id}/buddies/{userId}:
    delete:
      description: 'Untags a buddy from a checkin. Only the author can untag, and
        only while the checkin is still pending: once it is approved the workout is
        already theirs.'
      parameters:
      - description: checkin id
        in: path
        name: id
        required: true
        type: string
      - description: tagged user id
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove a tagged buddy
      tags:
      - checkins
...
  /me:
    get:
      description: Returns the authenticated user profile with achievements.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.MeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my profile
      tags:
      - me
    patch:
      consumes:
      - application/json
      description: Changes the theme and the status line. Every field is optional;
        only the ones present are written. Empty strings clear the status. The rank
        is derived from workouts and cannot be set.
      parameters:
      - description: fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.UpdateMeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update my profile
      tags:
      - me
...
  /rooms/{id}/freeze:
    delete:
      description: Cancels the scheduled freeze or unfreezes early. The cooldown counts
        the days actually used.
      parameters:
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel my freeze
      tags:
      - rooms
    post:
      consumes:
      - application/json
      description: 'Schedules a freeze: periods overlapping it are not judged. One
        active or scheduled freeze at a time; a cooldown of max(7, length) days follows
        each freeze.'
      parameters:
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      - description: freeze window, dates in YYYY-MM-DD
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.FreezeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Freeze'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Freeze my membership
      tags:
      - rooms
...
securityDefinitions:
  BearerAuth:
    description: 'JWT from POST /auth/telegram: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"